+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is used to list changes made in each version of the haproxy cookbook.
 - Fix ordering for `haproxy_listen`: `acl` directives should be applied before `http-request`
 - Add `hash_type` param to `haproxy_backend`, `haproxy_listen`, and `haproxy_config_defaults` resources
 - Add `reqirep` and `reqrep` params to `haproxy_backend`, `haproxy_listen`, and `haproxy_listen` resources
+- Add `sensitive` param to `haproxy_install`; set to false to show diff output during Chef run
 
 ## [v6.2.6](2018-11-05)